agregando clave para editar

+++ models/github/resolvers.ts
@@ -19,24 +19,30 @@ const resolversGithub ={
             return  codigoCreado;
         },
         
-        eliminarCodigoGithub: async(parent,args)=>{
-            if(Object.keys(args).includes('_id')){
-                const codigoEliminado=GithubModel.findOneAndDelete
-                    ({
-                    _id:args._id
-                    });
-                return codigoEliminado;
-            }  
+        eliminarCodigoGithub: async(parent,args,context)=>{
+            if (context.password===process.env.PASSWORD ){
+                console.log("la contraseña es correcta")
+                if(Object.keys(args).includes('_id')){
+                    const codigoEliminado=GithubModel.findOneAndDelete
+                        ({
+                        _id:args._id
+                        });
+                    return codigoEliminado;
+                }  
+            }
         },
            
-        editarCodigoGithub:  async(parent,args)=>{
-            const codigoEditado= await GithubModel.findByIdAndUpdate(args._id,{
-                descripcion:args.descripcion,
-                codigo:args.codigo,        
-            },
-                {new:true}  
-            );
-            return codigoEditado;     
+        editarCodigoGithub:  async(parent,args,context)=>{
+            if (context.password===process.env.PASSWORD ){
+                console.log("la contraseña es correcta")
+                const codigoEditado= await GithubModel.findByIdAndUpdate(args._id,{
+                    descripcion:args.descripcion,
+                    codigo:args.codigo,        
+                },
+                    {new:true}  
+                );
+                return codigoEditado;   
+            }  
         },
         
     }
